meta-ampere: support software revision using Ampere' format
OpenBMC uses os-release.bbappend to control the FW version, which is
the output of 'git describe --dirty --long'. The format of this command's
output is: {annotated-tag-name}-{num-commits-to-tag}-{current-commit-hash}.
'-dirty' will be appended to the output if the working space is not clean.

In addition, git describe --dirty --long only find an annotated tag, not
lightweight tag which will miss internal release.

This commit updates the rule to extract the right revision from Ampere
build and release tag.

Tested: tag with v1.02.100-ampere and check WebUI display revision
1.02.100.

Signed-off-by: Thu Cao <thcao@os.amperecomputing.com>