meta-ampere: mtjade: add POWERON option for the host sensors
Option POWERON_<type><index> is used to identify the host sensors.
They can be the virtual sensors of the host or the sensors which use
the same power domains with the host.
This option should be set to 1 for these kind of sensors.
Base on this option, phosphor-hwmon will activated or deactivated the sensors
when the host is on or off.
This commit adds the POWERON_<type><index> option for the host sensors in
mtjade platform.