meta-ampere: mtjade: Change to flash UEFI from offset 0x400000
The Host UEFI firmware layout is changed starting from 1.04 firmware.
The firmware image is generated without first 4MB data so it needs to be
flashed from offset 0x400000.
This commit changes to flash from offset 0x400000 to be compatible with
Host UEFI firmware 1.04 and later.

Note that this will not be compatible with firmware 1.03 and earlier
anymore.

Tested:
- Flash UEFI without 4Mb padding
- Check if UEFI boots successfully

Signed-off-by: Tung Nguyen <tung.nguyen@amperecomputing.com>