meta-ampere:mtjade:ras support record error overflow events
Due to the changes of host side, bit 8 of each RAS error counter register
will show the status of "error overflow" info. BMC will use this info
to record the SEL for this info. Be noted that, this bit will be cleared
when all error are handled (cleared) by BMC.
Error monitoring APIs updated to reflect those changing, the application should
also be changed. This commit supports that.

Testing:
	1. Issue a lot of RAS error (current buffer for each error is 8).
	2. Check the SEL list.
	1 | 01/19/2021 | 03:46:42 | OEM record c0 | 00cd3a | 078bffff0000 | RAS Processor CE_CPU_IError, errortype:Error Count is Overflow, subtype:0xff - Unknown_CPU_SubType(0xff), Socket0, CPM#:0x0
	a | 01/19/2021 | 03:46:43 | OEM record c0 | 00cd3a | 0c97ffff0000 | RAS Memory CE_Memory_IError, errortype:Error Count is Overflow, subtype:0xff, Socket0, instance:0x0

Signed-off-by: Thu B Nguyen <tbnguyen@amperecomputing.com>