meta-ampere: redfish: Fix the SpeedMbps=0 in EthernetInterfaces schema
The SpeedMbds of ethernet interface is implemented in getInterfaceInfo() from
the phosphor-networkd repo. It is disabled by NIC_SUPPORTS_ETHTOOL conditional
compilation. The default compile mode is to omit the getInterfaceInfo() which
causes the SpeedMbps not set.
This commit supports enable the NIC_SUPPORTS_ETHTOOL to report correct speed.

Tested:
1.Get SpeedMbps via rest:
$ curl -XGET --user root:0penBmc -H "Content-Type: application/json" \
       -H "If-match: *" --insecure \
       https://10.38.14.2/redfish/v1/Managers/bmc/EthernetInterfaces/eth1 \
       grep | "SpeedMbps"
2.Get speed of eht1 interface on BMC console
$ cat /sys/class/net/eth1/speed
The value getting from #1 and #2 match together.

Signed-off-by: Hieu Huynh <hieu.huynh@amperecomputing.com>