meta-ampere: sdr: Remove HOSTON_ configuration for the host sensors
With the changing in SMPro driver, the hwmon sysfs will not be bound
until the Host ready. This causes the Host sensor only available when
the Host is ON. Consequently, HOSTON_ attribute to set these sensors
active/inactive is not neccessary anymore.
Remove HOSTON_ in these sensors as they are redundant now.

Tested:
1. Boot up BMC with the host is Off.
2. Check the host sensors. They should be invisible.
3. Boot up BMC with the host is On.
4. Check the host sensors. They should be visible.

Signed-off-by: Thu B Nguyen <tbnguyen@amperecomputing.com>