meta-ampere: Support bind/unbind SMPro-hwmon driver use driver-binder
When BMC boots up with the host is Off, linux kernel will load the SMPro
drivers, but the devices are not available so it will not bind the drivers.
Moreover, linux kernel don't support auto bind the driver when device is ready.
So the BMC application should bind/unbind the drivers base on the host status.

Driver-binder in ampere-misc is designed for that purpose.
This commit adds support the binding/unbinding SMPro hwmon driver
using Driver-binder.

Tested:
        1. Boot up BMC with the host is Off.
        2. There are no host sensors.
        3. Turn on the host.
        4. The host sensors should be available.

Signed-off-by: Thu B Nguyen <tbnguyen@amperecomputing.com>