meta-ampere: sdr: update expression to calculate S*_DIMMG0_Temp
S*_DIMMG0_Temp sensors are reported by phosphor-virtual-sensor which use
Exprtk lib to calculate values for the sensors.
There is an issue in Exprtk lib which returns na for the max function
when the first item is na.
Add 0 to the first item for the max() function to workaround the issue.

Tested:
	1. Remove DIMM at channel 0, install it to channel 1.
	2. Make sure S*_DIMMG0_Temp is max temperature of the DIMM temperature
	   from channel 0 to 3.

Signed-off-by: Thu Ba Nguyen <tbnguyen@amperecomputing.com>