meta-ampere: switch the SPI0_PROGRAM_SEL in powering system on/off
When powering ON the Host, no code to switch the SPI0_PROGRAM_SEL signal
to Host. This makes the Host fail to boot UEFI if users manually switch
the Host SPI-NOR to BMC.
In addition, when the Host is OFF, switching the SPI-NOR to Host might
cause leakage current.
This commit controls the SPI0_PROGRAM_SEL GPIO signal to switch the Host
SPI-NOR to Host when powering ON and switch to BMC when powering OFF.

Tested:
1. Manually set GPIO 226=0. Powering the Host, ensure Host can boot.
2. Turn the Host OFF, check if BMC can access the Host SPI-NOR.

Signed-off-by: Thang Q. Nguyen <thang@os.amperecomputing.com>