meta-ampere: dbus-sensor: Support devName option to match ADC sensors
AdcSensor daemon in dbus-sensor does not support iio-hwmon-mux and it
only supports one iio-hwmon device with the matched name is iio-hwmon.
This doesn't match with Mt.Jade platform which have tree iio-hwmon-mux
and three iio-hwmon devices in DTS.

These patches improve AdcSensor daemon to support multi iio-hwmon devices.
And use the setting of devName in the sensors configuration to match
the ADC sensors.

Tested:
	1. Check the ADC sensor in xyz.openbmc_project.ADCSensor dbus.
	2. All configured ADC sensors should be listed.

Signed-off-by: Thu Ba Nguyen <tbnguyen@amperecomputing.com>