meta-ampere: entity-manager: Support parsing Mt.Jade motherboard EEPROM
FruDevice of entity-manager is failed to parse Mt.Jade EEPROM with error
"Fru area offsets are not in required order as per Section 17 of Fru
specification". This cause entity-manager can't detect Mt.Jade platform to
load sensor configuration.
Because this Fru format is used by the Ampere customer so this error should
be bypassed as a workaround solution. This makes sure Mt.Jade platform can be
detected and sensor model is loaded successfully.
This commit support that workaround solution.

Tested:
	1. Call "busctl tree xyz.openbmc_project.FruDevice".
	2. Mt_Jade_Motherboard should be listed.

Signed-off-by: Thu Ba Nguyen <tbnguyen@amperecomputing.com>