meta-ampere: meta-jade: sdr: Increase the Multiple value of S0/S1_Core_VRD_Curr
The current Multiple of S0/S1_Core_VRD_Curr sensors are 80. So the values of
these sensors are from 0x08 to 20.4A. But the comsumed currents of the Core VRD
are bigger than 20.4. The observed value is 25A, it exceeds the value range
of these sensors. This commit changes Multiplex to 0xA0 so the value range is
from 0x16 to 40.8.

Tested:
	1. Check the values of S0/S1_Core_VRD_Curr sensors inBMC Web and
	ipmitool sensor list. They should be similar.

Signed-off-by: Thu B Nguyen <tbnguyen@amperecomputing.com>