meta-ampere: nvme: check for change before updating inventory data
NVMe data for NVMe devices are read continuously. And the inventory
data is set to dbus without checking if any change or not.
This causes the dbus too busy that sometime hang because of timeout.

This patch will add the serial number, smartwarning, statusflag checking
before set the inventory properties. Inventory data is just be updated
if anything changed.

Tested: Read the dbus sensor and inventory without delay or timeout
1. Check inventory information at the object Dbus
/xyz/openbmc_project/inventory/system/chassis/motherboard/nvme(index)
2. Check sensor temperature at the object Dbus
/xyz/openbmc_project/sensors/temperature/nvme(index)

Signed-off-by: Chanh Nguyen <chanh@os.amperecomputing.com>