meta-ampere: dbus-sensors: Handle errors when smpro-hwmon drivers are unbinded
When the host is off, the smpro-hwmon drivers are unbinded, the host sensor
paths are not available to read. The value property of the host sensor dbus
interface should be updated to NaN.
This commit update the socsensor daemon to keep reading the host sensors until
the host sensor paths are available again.

Tested:
	1. Boot up the BMC with the host is on.
	2. Power off the host.
	3. The host sensor dbus interface should be available.
	4. The values of the host sensors should be NaN.
	5. The values of the virtual sensors should be 0.

Signed-off-by: Thu B Nguyen <tbnguyen@amperecomputing.com>