meta-ampere: ampere-hostctrl: Only remove /run/openbmc/* when power on
When the BMC boot up with the host is Running, /run/openbmc/chassis@0-on is
created and obmc-chassis-poweron@.target is started. It then calls
ampere-chassis-poweron.service. But the /run/openbmc/chassis@0-on is existing,
there is no need to actually turn power on. ampere-chassis-poweron don't
power on the host in this case. It should not remove /run/openbmc/chassis@0-on
and /run/openbmc/host@0-on.

Tested:
	1. Boot up BMC with the host is On.
	2. The host state should be Running.

Signed-off-by: Thu B Nguyen <tbnguyen@amperecomputing.com>