meta-ampere: ampere-host-check: Verify the host state in power on
When power on the host, phosphor-state-manager does not use the software
interfaces to verify the host state, it is assuming that the host is running
after all of required services of the power on target service were started.
This causes the CurrentHostState dbus property is Running before the host is
ready.
In Mt.Jade platform, the SCP drives S0_FW_BOOT_OK to high when the host
is already on. This GPIO can be used to identify the host state. One
service should be added to the required list. This service checks the
host state in power on and triggers obmc-host-quiesce@%i.target when the host
is failed to boot.
ampere-host-check package adds ampere-host-on-host-check@.service for that
purpose.

Tested:
	1. Boot up BMC with the host is off/on. Verify CurrentHostState.
	2. Call Ipmitool power off/on. Verify the CurrentHostState.
	3. Call Ipmitool chassis power off/on. Verify the CurrentHostState.
	4. Call Ipmitool power cycle. Verify the CurrentHostState.
	5. Call Ipmitool chassis power cycle. Verify the CurrentHostState.
	6. Power off the host. Set power policty to Always-on. AC power.
	The host should be on. And CurrentHostState should be Running.
	7. Power on the host. Set power policty to Always-off. AC power.
	The host should be off. And CurrentHostState should be Off.

Signed-off-by: Thu B Nguyen <tbnguyen@amperecomputing.com>