meta-ampere: implement Altra Boot Progress handling
Implement the Boot Progress handling in a new design which purely using
OpenBMC framework. The code to handle boot progress is written in bash
shell so that it is easy to upstream the feature.

BZ: https://den-bugzilla.amperecomputing.com/show_bug.cgi?id=79640

Tested: check boot progress is updated in Redfish for:
1. Reboot Host.
2. IPMI chassis power reset
3. IPMI chassis power cycle
4. IPMI chassis power off, then on
5. WebUI Orderly – OS shuts down, then server reboots
6. WebUI Immediate – Server reboots without OS shutting down

Signed-off-by: Thang Q. Nguyen <thang@os.amperecomputing.com>