meta-ampere: sesponse the thresholds sensors with some un-set thresholds
The sensor packages such as dbus-sensors and phosphor-hwmon packages supports
creating the dbus interface for the sensors which have some the un-set
thresholds.
But phosphor-ipmi-host requires the sensors have enough "WarningLow",
"WarningHigh", "CriticalLow" and "CriticalHigh" thresholds. It doesn't
response the sensor thresholds to ipmitool request if the sensors miss any
kind of these thresholds. This commit removes that limitation.

Tested:
        1. Call "ipmitool sensor list".
        2. Check the thresholds of the sensors.

Signed-off-by: Thu B Nguyen <tbnguyen@amperecomputing.com>