meta-ampere: fwupdate: use "obmcutil poweron" to poweron the Host
Phosphor-state-manager only turns on the power when it responses for
power action "chassisOn" or "forceOn". It does not transit the host
state. Using "obmcutil chassison" to turn on the host is uncorrected.
This commit corrects that action by using "obmcutil poweron" to turn
on the host in firmware update scripts.

Tested:
    1. Do SCP/UEFI firmware update.
    2. The host state should be Running after firmware update.

Signed-off-by: Thu B Nguyen <tbnguyen@amperecomputing.com>