meta-ampere: Fix IPMI firmware revision report
Ampere OpenBMC tags the firmware revision in format
vX.YY.YYYYMMDD-ampere for internal testing build and vX.YY.Z00 for
external release where X and YY are decimal number.
However, phosphor-host-ipmid translates it as hexa-decimal number so the
firmware revision in the "ipmitool mc info" is wrong.
This commit adds a patch to fix the issue

Tested:
1. Tag the firmware revision as v1.10.20210817-ampere and check if
   ipmitool mc info returns 1.10 for firmware revision.
2. Tag the firmware revision as v1.10.100-ampere and check if
   ipmitool mc info returns 1.10 for firmware revision.

Signed-off-by: Thang Q. Nguyen <thang@os.amperecomputing.com>