meta-ampere: add driver-binder to bind smpro driver
This commit replaces the external driver-binder package in
ampere-platform-mgmt by the internal ampere-driver-binder package.
This package will bind SMpro-hwmon drivers at the first power on of
the host. It check the existing of /run/openbmc/host@0-on which is
created by ampere-host-check to indentity the host state.

The SMPro drivers names and paths are configured in drivers-conf.sh.

Tested:
    1. Boot up the BMC with the host is off.
    2. There are no SMPro-hwmon drivers in
       /sys/bus/platform/drivers/smpro-hwmon.
    3. Power on the host.
    4. The SMPro-hwmon drivers should be there.

Signed-off-by: Thu B Nguyen <tbnguyen@amperecomputing.com>