meta-ampere: Fix the ipmitool lan print 3 does not work
The purpose of the commit 0a2abc8 in the phosphor-host-ipmid repo is
compares the LAN interfaces defined in the channel_config file with
the LAN devices reported by the getifaddrs. If a LAN channel is
described in the channel_config file but missing from the getifaddrs
device list, the channel is marked as invalid.
Currently, the getifaddrs does not list the usb0 interface in our
system. It causes the issue the ipmitool lan print 3 does not work.
This reverts commit 0a2abc803ffb4e01d10d4c9fb78475c3012df496 from the
phosphor-host-ipmid repo as temporary workaround to fix this issue.

Tested:
1. Get the USB ethernet: # ipmitool lan print 3
   The USB ethernet will show the information.

Signed-off-by: Hieu Huynh <hieuh@os.amperecomputing.com>