meta-ampere: mtjade: fix sensor binding fail
After BMC is rebooted while the Host is ON, the /run/openbmc/host@0-on
file is created but no one delete it. This makes the next power
cycle/reset works improperly as several services think the Host is
currently ON but actually it is not.
This commit fixes the issue by deleting the file before power OFF or
reset the Host.

Tested: check "ipmitool sdr list" works correctly in the following
cases:
1. IPMI chassis power cycle
2. IPMI chassis power reset
3. IPMI chassis power off, then on.
4. Execute poweroff command from Host, then turn on via IPMI.
5. Execute reboot command from Host.

Signed-off-by: Thang Q. Nguyen <thang@os.amperecomputing.com>