meta-ampere: Disable host control in phosphor-fan
With json configuration format, phosphor-fan-monitor will shutdown the
host when the fan tachometer sensors are not present because it assumes
there is no other fan control service.

In Ampere systems, the fans will be set to run at default speed in early boot.
So shutdown the host is unnecessary. This commit supports option to disable host
control in phosphor-fan.

Tested:
1. Stress BMC reset cold 200 times.
2. The host state should not be changed.

Signed-off-by: Thu Ba Nguyen <tbnguyen@amperecomputing.com>