meta-ampere:mtjade: fix Host Firmware update intermittent failure
For Host FW update, the ampere_flash_bios.sh script will check if the
HNOR partition is not available, then bind the Aspeed SMC driver. The
issue is in case Host is ON, the HNOR partition is not available but
the Host SPI '1e630000.spi' exists in the driver directory. This causes
the bind command to fail.

This commit supports unbind if HOST SPI exists to fix the issue.

Signed-off-by: Thang Q. Nguyen <thang@os.amperecomputing.com>