meta-ampere:mtmitchell: reduce fan present restart frequency
When BMC boots up with the host is on, sometimes phosphor-fan-presence-tach
starts early before power services, and it keeps restarting every 1 second
because there is no fan tach sensors yet. This causes the failure of
obmc-chassis-on service which leads to the failure in updating the host status.
This commit reduces the restart frequency of phosphor-fan-presence-tach service
to prevent the failure of obmc-chassis-on.

Tested:
1. Stress bmc reset cold.
2. The host state should be updated correctly.

Signed-off-by: Thu Ba Nguyen <tbnguyen@amperecomputing.com>