meta-ampere: state-manager: allow power cycle when reboot count is 0
Currently, when the user requests to power cycle the host, the software
shall check the reboot-counter to decide whether to reboot the host or
not. This makes "ipmitool chassis power cycle" command failed if
reboot-counter is 0.

This commit overrides phosphor-reboot-host@.service, this service shall
call obmc-host-startmin@%i.target even reboot count is 0.

Tested:
    1. Set AttemptsLeft to 0.
       busctl set-property xyz.openbmc_project.State.Host \
       /xyz/openbmc_project/state/host0 \
       xyz.openbmc_project.Control.Boot.RebootAttempts \
       AttemptsLeft u 0
    2. Request power cycle to the host.
       ipmitool chassis power cycle
    3. Host is rebooted.

Signed-off-by: Thang Tran <thuutran@amperecomputing.com>