meta-ampere: buttons: avoid get service fail
Current code to check if the platform is multi-host is via
/xyz/openbmc_project/Chassis/Buttons/HostSelector which always returns
true. The root cause comes from the getService() function not checking
whether the multi-host check is successful or not.

This commit adds try-catch to handle the case the dbus call fails that
returns a null string to indicate it is single host platform.

Tested:
Host CPU on
1. Press Power Button ->  Check CPU host off|on
2. Press Reset Button ->  Check CPU host reset

Signed-off-by: HuyLe <hule@amperecomputing.com>