meta-ampere: host-ipmid: replace static convert with rounding
Issue: When the users request to get the sensor thresholds via "ipmitool
sensor get <sensor_name>" command, sometimes the value is not correct.
e.g: The real sensor threshold is 2.01, the "scale" = 0, "exponentR" =
-3 and "coefficientM" = 10. . After calculating, the expect return value
is 201, but the actual result is 200 (ipmitool displays the threshold's
value is 2.00).

Root cause: The "float point" cannot present exactly the "double"
value in some cases. Therefore, after calculating, the value of the sensor
threshold is incorrect when converting from double to uint8_t.

Solution: Replace static converting with rounding.

Tested:
    1. Check the result of the threshold whose vaule is 2.01
       ipmitool sensor get <2.01_Sensor_name>
    2. The threshold value is 2.01

Signed-off-by: Thang Tran <thuutran@amperecomputing.com>