meta-ampere: support ampere power/host already on targets
There is no target to identity power state that is "already on"
and host state of "already running". This commit adds those targets.
The targets will be the check prior to the start of ampere services such as
drivers binder, hang detection...
obmc-power-already-on@.target will always activate when
chassis-power-control is on. It will be deactivated before triggering
chassis-power-control to off.
obmc-host-already-on@.target will be activated when the host%i-ready
is triggered. The target will be deactivated before turning
off/reseting the host.

Tested:
1. The targets should be activated/deactivated per design

Signed-off-by: Thu Nguyen <thu@os.amperecomputing.com>